requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
colorama==0.4.4
Django==2.2.20
django-grappelli==2.14.3
iniconfig==1.1.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.3
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
